vm: Remove 'apiStack' hack.
The 'apiStack' hack is an attempt to reduce the number of indirections
to access the fiber/thread 'stack_base'.

Incidently, it also serve as a flag to indicate if we are currently in
an C API context.

So this patch conserve the flag behaviour, by adding the 'is_api_call'
boolean, while adding 'stack_base' to the fiber/thread.

While 'stack_base' is not garanteed to be up to date from where ever in
inside the vm, it must always be at its correct place when acceced from
user C code.